Timer in Input Capture mode
In this Lab we will compute a signal input frequency using the Timer's IC.In order to represent this functionality we will proceed as follows:

-Generate a PWM Signal using Timer 3 [Channel 1] at PB4 Pin (D5) .
-Captuer two consuctive rising pulses using the Timer 2 [Channel 1] at PA5 Pin (D13).

PWM Signal frequency = 1 us.

Refer to the CubeMX configuration to see the timers parameters.

TIM_IC_CaptureCalleback(htim) Function definition, equations and different Steps are explained as comments in the Cube IDE.

Enjoy !