ADC in polling mode
Set up a new project as usual with system clock @ 72MHz
Set up An Analog Input Pin (Channel 7) In single Conversion Mode (The Pot. Pin)
Set up timer2 in PWM mode with output on channel 1 (The LED Pin)
The Application will have 3 versions each does the same thing which is read the ADC result and move it to the timer CCR register which decides the PWM duty cycle percentage on the output LED pin.

*Example 1, ADC is used in blocking mode (polling)

Example 2, ADC is used in non-blocking mode (interrupt)

Example 3, ADC is used in non-blocking mode (DMA)