15_SPI Communication MCU-MCU
15_SPI Master Transmitter  in polling Mode: Configured as an "only Transmitter Master" operation mode in the CubeMX.
15_SPI Receiver Slave in Polling Mode: Configured as an "only Receiver Slave" operation mode in the CubeMX.

The Transmiter send the same data every 1s with a timeout of 1s (The TX_data must be sent before that time).
In the other hand the receiver is reading the data every 0.5s with a 1s timeout (RX_buffer must take a maximum of the fixed timeout to be received).

If the timeout is not respected in one of the configurations an error will be triggered and the programme will move to the Error_Handler function.