17_DAC on single mode
Here is the link for the lab: https://deepbluembedded.com/stm32-dac-tutorial-example-hal-code-analog-signal-genreation/

Instead of monitoring the output voltage using a oscilloscope. I have introduced a variable that stores the ODR1 of the DAC1. Then In order to see the result (array values been transferred to the DAC_out) just by debugging  and opening the live expression u can see the values being transferred.