Mini Project interfacing STM32 with Smartphone via Blutooth HC06
Using Arduino controller: When pressing labeled buttons on the smartphone we toggle the LEDs as the following:
-Green LED -> UART value = '1' -> Green LED set on.
-Blue LED -> UART value = '2' -> Blue LED set on.
-Yellow LED -> UART value = '3' -> Yellow LED set on.
we can use the terminal on the smartphone interface to send row data.
In parallel am using an ADC entry from a potentiometer to control white LED  Dimmer trough the MAP function (to match ADC and PWM resolutions).